+++ app/Providers/AppServiceProvider.php
@@ -4,10 +4,10 @@ namespace App\Providers;
 
 use App\Repositories\Category\CategoryRepository;
 use App\Repositories\Category\CategoryRepositoryInterface;
-use App\Repositories\Eloquent\Document\DocumentRepository;
-use App\Repositories\Eloquent\Document\DocumentRepositoryInterface;
 use App\Repositories\User\UserRepository;
 use App\Repositories\User\UserRepositoryInterface;
+use App\Repositories\Document\DocumentRepository;
+use App\Repositories\Document\DocumentRepositoryInterface;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -30,14 +30,6 @@ class AppServiceProvider extends ServiceProvider
         $this->app->singleton(
             DocumentRepositoryInterface::class,
             DocumentRepository::class
-        );
-        $this->app->singleton(
-            \App\Repositories\Eloquent\User\UserRepositoryInterface::class,
-            \App\Repositories\Eloquent\User\UserRepository::class
-        );
-        $this->app->singleton(
-            \App\Repositories\Eloquent\Category\CategoryRepositoryInterface::class,
-            \App\Repositories\Eloquent\Category\CategoryRepository::class
         );
     }
