+++ app/Repositories/CategoryRepository/CategoryRepositoryInterface.php
@@ -8,11 +8,13 @@ interface CategoryRepositoryInterface extends RepositoryInterface
 {
     public function getCategoriesRoot();
 
-    public function getChildCategories();
-
     public function countChildCategories($category);
 
     public function updateChildCategoriesParent($category, $attribute = []);
 
     public function updateDocumentsCategory($category, $attribute = []);
+
+    public function getChildCategoriesFromRoot();
+
+    public function getDocumentsByCategory($id);
 }
