--- app/Repositories/Eloquent/Category/CategoryRepositoryInterface.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App\Repositories\Eloquent\Category;
-
-interface CategoryRepositoryInterface
-{
-    /**
-     * @return mixed
-     */
-    public function getChildCategoriesFromRoot();
-
-    /**
-     * @param $id
-     * @return mixed
-     */
-    public function getDocumentsByCategory($id);
-}
