--- app/Repositories/EloquentRepository.php
@@ -1,111 +1,0 @@
-<?php
-
-namespace App\Repositories;
-
-use Illuminate\Contracts\Container\BindingResolutionException;
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Database\Eloquent\Model;
-
-abstract class EloquentRepository implements EloquentRepositoryInterface
-{
-    /**
-     * @var Model
-     */
-    protected $model;
-
-    /**
-     * EloquentRepository constructor.
-     * @throws BindingResolutionException
-     */
-    public function __construct()
-    {
-        $this->setModel();
-    }
-
-    /**
-     * get model
-     * @return string
-     */
-    abstract public function getModel();
-
-    /**
-     * Set model
-     * @throws BindingResolutionException
-     */
-    public function setModel()
-    {
-        $this->model = app()->make(
-            $this->getModel()
-        );
-    }
-
-    /**
-     * Get All
-     * @return Collection|static[]
-     */
-    public function all()
-    {
-
-        return $this->model->all();
-    }
-
-    /**
-     * Get one
-     * @param $id
-     * @return mixed
-     */
-    public function find($id)
-    {
-        try {
-            $result = $this->model->findOrFail($id);
-        } catch (\Exception $e) {
-            return false;
-        }
-        return $result;
-    }
-
-    /**
-     * Create
-     * @param array $attributes
-     * @return mixed
-     */
-    public function create(array $attributes)
-    {
-        return $this->model->create($attributes);
-    }
-
-    /**
-     * Update
-     * @param $id
-     * @param array $attributes
-     * @return bool|mixed
-     */
-    public function update($id, array $attributes)
-    {
-        $result = $this->find($id);
-        if ($result) {
-            $result->update($attributes);
-            return $result;
-        }
-
-        return false;
-    }
-
-    /**
-     * Delete
-     *
-     * @param $id
-     * @return bool
-     */
-    public function delete($id)
-    {
-        $result = $this->find($id);
-        if ($result) {
-            $result->delete();
-
-            return true;
-        }
-
-        return false;
-    }
-}
