--- app/Repositories/EloquentRepositoryInterface.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace App\Repositories;
-
-interface EloquentRepositoryInterface
-{
-    /**
-     * Get all
-     * @return mixed
-     */
-    public function all();
-
-    /**
-     * Get one
-     * @param $id
-     * @return mixed
-     */
-    public function find($id);
-
-    /**
-     * Create
-     * @param array $attributes
-     * @return mixed
-     */
-    public function create(array $attributes);
-
-    /**
-     * Update
-     * @param $id
-     * @param array $attributes
-     * @return mixed
-     */
-    public function update($id, array $attributes);
-
-    /**
-     * Delete
-     * @param $id
-     * @return mixed
-     */
-    public function delete($id);
-}
