+++ tests/Unit/Http/Controllers/DocumentControllerTest.php
@@ -8,9 +8,9 @@ use App\Http\Requests\DocumentRequest;
 use App\Models\Category;
 use App\Models\Document;
 use App\Models\User;
-use App\Repositories\Eloquent\Category\CategoryRepositoryInterface;
-use App\Repositories\Eloquent\Document\DocumentRepositoryInterface;
-use App\Repositories\Eloquent\User\UserRepositoryInterface;
+use App\Repositories\Category\CategoryRepositoryInterface;
+use App\Repositories\Document\DocumentRepositoryInterface;
+use App\Repositories\User\UserRepositoryInterface;
 use Illuminate\Http\Request;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Collection;
